Move Binance and CoinbasePro into the package exchange
