+++ troubleshoot.py
@@ -1,10 +1,5 @@
-import numpy as np
-import matplotlib.pyplot as plt
-
 from models.PyCryptoBot import PyCryptoBot
 from models.Trading import TechnicalAnalysis
-from models.Binance import AuthAPI as BAuthAPI, PublicAPI as BPublicAPI
-from models.CoinbasePro import AuthAPI as CBAuthAPI, PublicAPI as CBPublicAPI
 from views.TradingGraphs import TradingGraphs
 
 app = PyCryptoBot()
